+++ sdk/storage/azure-storage-blob/tests/test_largest_block_blob.py
@@ -17,7 +17,7 @@ from azure.storage.blob import (
     BlobServiceClient,
     BlobBlock
 )
-from azure.storage.blob._shared.base_client import format_shared_key_credential
+from azure.storage.blob._shared.base_client import _format_shared_key_credential
 
 from _shared.testcase import StorageTestCase, GlobalStorageAccountPreparer
 
@@ -97,7 +97,7 @@ class StorageLargestBlockBlobTest(StorageTestCase):
     @GlobalStorageAccountPreparer()
     def test_put_block_bytes_largest_without_network(self, resource_group, location, storage_account, storage_account_key):
         payload_dropping_policy = PayloadDroppingPolicy()
-        credential_policy = format_shared_key_credential([storage_account.name, "dummy"], storage_account_key)
+        credential_policy = _format_shared_key_credential(storage_account.name, storage_account_key)
         self._setup(storage_account, storage_account_key, [payload_dropping_policy, credential_policy])
         blob = self._create_blob()
 
@@ -157,7 +157,7 @@ class StorageLargestBlockBlobTest(StorageTestCase):
     @GlobalStorageAccountPreparer()
     def test_put_block_stream_largest_without_network(self, resource_group, location, storage_account, storage_account_key):
         payload_dropping_policy = PayloadDroppingPolicy()
-        credential_policy = format_shared_key_credential([storage_account.name, "dummy"], storage_account_key)
+        credential_policy = _format_shared_key_credential(storage_account.name, storage_account_key)
         self._setup(storage_account, storage_account_key, [payload_dropping_policy, credential_policy])
         blob = self._create_blob()
 
@@ -211,7 +211,7 @@ class StorageLargestBlockBlobTest(StorageTestCase):
     @GlobalStorageAccountPreparer()
     def test_create_largest_blob_from_path_without_network(self, resource_group, location, storage_account, storage_account_key):
         payload_dropping_policy = PayloadDroppingPolicy()
-        credential_policy = format_shared_key_credential([storage_account.name, "dummy"], storage_account_key)
+        credential_policy = _format_shared_key_credential(storage_account.name, storage_account_key)
         self._setup(storage_account, storage_account_key, [payload_dropping_policy, credential_policy])
         blob_name = self._get_blob_reference()
         blob = self.bsc.get_blob_client(self.container_name, blob_name)
@@ -237,7 +237,7 @@ class StorageLargestBlockBlobTest(StorageTestCase):
     @GlobalStorageAccountPreparer()
     def test_create_largest_blob_from_stream_without_network(self, resource_group, location, storage_account, storage_account_key):
         payload_dropping_policy = PayloadDroppingPolicy()
-        credential_policy = format_shared_key_credential([storage_account.name, "dummy"], storage_account_key)
+        credential_policy = _format_shared_key_credential(storage_account.name, storage_account_key)
         self._setup(storage_account, storage_account_key, [payload_dropping_policy, credential_policy])
         blob_name = self._get_blob_reference()
         blob = self.bsc.get_blob_client(self.container_name, blob_name)
@@ -257,7 +257,7 @@ class StorageLargestBlockBlobTest(StorageTestCase):
     @GlobalStorageAccountPreparer()
     def test_create_largest_blob_from_stream_single_upload_without_network(self, resource_group, location, storage_account, storage_account_key):
         payload_dropping_policy = PayloadDroppingPolicy()
-        credential_policy = format_shared_key_credential([storage_account.name, "dummy"], storage_account_key)
+        credential_policy = _format_shared_key_credential(storage_account.name, storage_account_key)
         self._setup(storage_account, storage_account_key, [payload_dropping_policy, credential_policy],
                     max_single_put_size=LARGEST_SINGLE_UPLOAD_SIZE)
         blob_name = self._get_blob_reference()